Add files via upload
TO-DO LIST
This is the TASK-2 given by the octanet services.
Objective: The goal of this task is to create a functional and user-friendly To-Do List web page. The page should allow users to add, delete, and mark tasks as complete It should also have the ability to prioritize tasks, organize them into categories, and set due dates.

Requirements: Design a clean and intiative user interface for the To-Do List web page. Include features such as task entry field, buttons for adding, deleting, and marking tasks as complete web page.